Fix #4408, fix #4407, fix #4405 - Fix Now-Playing, title syncing with CarPlay, Error Message on iOS 15 (#4409)
Allow playlist to properly handle cancelling of requests because on iOS 15, table cells are laid out asynchronously and causes a race-condition between playback and layout. Since layout cancels requests, an error shows.
Fixed not being able to delete an item before it finishes loading on a very very slow network.