Fix BatAdsAdNotificationServingTest.ServeAdWithAdServingVersion2 setup
This test was not being properly setup since a string was expected to
be set as the value for ad_serving_parameters["ad_serving_version"],
but instead an integer (2, instead of "2") was being passed.

And we didn't noticed until now because the parsing problem would go
unnoticed as a DLOG(WARNING), resulting in the default value being
used, but now that the DLOG(WARNING) became a NOTREACHED() the test
is crashing instead.

To fix this, we just make sure we actually set a string as the value
for ad_serving_parameters["ad_serving_version"], and also add a couple
of checks more in the test suite to make sure that AdServing is enabled
and that the version being used is, indeed, version 2.

Chromium change:

https://chromium.googlesource.com/chromium/src/+/bf076ce0d82267340f9ba6aa4849d8a37aa8a0db

commit bf076ce0d82267340f9ba6aa4849d8a37aa8a0db
Author: Christian Dullweber <dullweber@chromium.org>
Date:   Wed Oct 20 10:35:26 2021 +0000

    AccuracyTips: Fix field_trial config and add DCHECKs to config parser

    FeatureParam<TimeDelta> does not support days as parameter value, so our
    "1d" config is not parsed correctly and instead 7 days are used.
    Fix the config by using hours and throw errors when an incorrect config
    is parsed instead of just logging a warning to avoid future issues.

    Bug: 1261162