Use /etc/brave/policies even if it doesn't exist.
PathService::Override() requires that the new target path exists,
and tries to create it if it doesn't. Since the user doesn't have
permission to create `/etc/brave` the override fails if an
administrator hadn't already installed something there, and the
original `/etc/chromium` path for policy files remains active.

Current uses of this path are few and seem to be robust against
the directory not existing, so this seems a safer fix than patching
the upstream default.

Closes brave/brave-browser#19052