P3A: Test Rewards.EnabledDuration measurement.
Turn Brave Rewards on and off, checking for the appropriate
measurements. Modify the recorded timestamp to mock the passage
of time.

According to `RewardsServiceImpl::IsRewardsEnabled`, rewards
are enabled if either of Ads or AutoContribute is enabled.
In general switching these may call `OnPreferenceChanged`
asynchronously after a round-trip through the ledger process.

This commit overrides `RewardsObserver::OnAdsEnabled` to keep
track of the enabled state, and adds a `WaitForRewardsEnable`
method which blocks until changes have propagated so the test
can wait for this before checking for the expected histogram
values.

It adds a similar invocation to `WaitForRewardsInitialization`.
According to Emerick Rogul this was the intended use of the
corresponding RunLoop data member. Likely it wasn't necessary
in practice because the existing tests all call `FetchBalance`
which calls its own local RunLoop.

Rather than add a parallel `RewardsObserver::OnAutocontributeEnabled`
callback, the new duration tests only check the ads case.

I had further problems with multiple enable/disable cycles; it seems
something with the ledger shutdown code wasn't running to completion
even with an extra runloop or delay. To work around this, each
duration check is written as a separate test.

Also, prefer assigning a `make_unique` to `reset` for unique_ptr
data members per style lint.