Disable navigator.keyboard API when fingerprinting protection is on.
Fixes brave/brave-browser#3964

Have not found a convenient way to return navigator.keyboard as
undefined yet, so for now making it a null.

Since this API is not supported by all browsers making an assumption
that web developers would check the API availability before using it
in the form of `if (navigator.keyboard)...`, in which case a null
should work just as well.