Pull Request template encourages squashing of fixup / review feedback commits
Encouraging a clean git history makes for a more readable log of file changes. 
Commit message are much more readable when they talk about features or bug fixes (as described in the existing checklist item about good commit descriptions).

Aim is to prevent a lot of commits which talk about the journey of a PR, such as fixing CI build issues and addressing review feedback, rather than talking about improvements to the project.