Adapt Brave DarkMode browser tests to NavigationThreadingOptimizations
The way these tests check that what prefers-color-scheme reports is
dependant on BraveShields fingerprinting mode does not work anymore
with NavigationThreadingOptimizations enabled, as the page's title
won't be updated on time as expected, as it was done before.

Instead, we need to do what upstream's PrefersColorSchemeTest do
and inspect what prefers-color-scheme reports via JavaScript, so
that changes to the fingerprinting mode can be immediately checked
without relying on the page's title. Additionally, I found out that
we also need to call WebContents::OnWebPreferencesChanged() when
calling TestNativeTheme::SetDarkMode() (like upstream does already)
or tests that don't change the shields mode won't work either, so
this patch adds a SetDarkMode() method to the test class to make
sure that this is done whenever setting the mode on the theme.

Last, add a new PrefersColorSchemeWithDefaultFingerprinting() test
to check the color scheme mode via the page title once in a scenario
where the shields fingerprinting mode is left at its default value,
mimicking what the PrefersColorSchemeTest.PrefersColorScheme test
from upstream does as well.

PS: All the BraveDarkMode* tests tests pass regardless of whether the
NavigationThreadingOptimizations feature is enabled or disabled.

Chromium change:

https://github.com/chromium/chromium/commit/39e0e517ea133ca2a512c7daa98979a59b26bef2

commit 39e0e517ea133ca2a512c7daa98979a59b26bef2
Author: Clark DuVall <cduvall@chromium.org>
Date:   Tue Nov 30 22:25:04 2021 +0000

    Enable NavigationThreadingOptimizations by default

    This has been running at 10% stable and results are good.

    Bug: 1187753