Fix linking errors on Windows for npm run unit_tests
channel_override_work_item.{h,cc} are only included if is_chrome_branded
is set in the //chrome/installer/setup:lib target, which is a dependency
of the unit_tets target, causing a link issue due to our chromium_src
override in chromium_src/chrome/installer/setup/install_worker.cc, which
requires ChannelOverrideWorkItem to be enabled.

This is not a problem for Brave at the moment since we have a patch for
the //chrome/installer/setup:setup target that makes sure to add those
sources, but since that does not work for building the unit_tests target
as it references //chrome/installer/setup:lib instead.

The solution is to move the patch in //chrome/installer/setup/BUILD.gn
to add the sources in //chrome/installer/setup:lib, instead of doing it
in  //chrome/installer/setup:setup (which depends on :lib anyway).

Resolves https://github.com/brave/brave-browser/issues/20105