Dump the test's full output to the xUnit output xml file
Selectively picking up the affected file name, line number and error
message for each part of a tests' results means we won't be dumping
other bits into the output XML that are present on the standard output
when running the tests from the console, resulting in information loss.

More specifically, this becomes an bigger annoyance when a new failure
is detected by `npm run network-audit`, since the requests causing that
suite to fail will only be printed as a part of the test's full output
message (instead of being tied to each TestResultPart object), forcing
developers to check the full log in Jenkins to figure out what failed.

Therefore, let's stop iterating over the |test_result_parts| vector
from TestResults, and dump the contents of its |output_snippet| instead
to make sure we don't lose any bit of information that could be useful
to diagnose a test failure.

Resolves https://github.com/brave/brave-browser/issues/20111