addblock_rust: Remove unnecessary mem::forget from ffi wrapper.
These two functions call `std::mem::forget(ptr)` on a raw `c_char`
pointer. This is unnecessary; the previous `.into_raw()` has already
consumed ownership of the CString, so there's nothing to forget,
and returning the pointer correctly transfers ownership of the
underlying buffer to the caller. Put another way `*mut c_char`
is a `Copy` type and does not have a destructor, so `mem::forget`
is a no-op.

The way `engine_hidden_class_id_selectors()` instead returns the
`into_raw` value directly is correct.

Corrects a clippy warning.