Fix delta auto-update file generation on Windows
Chromium's courgette.exe runs into memory allocation errors when called on
Brave's binaries. Reducing the log page size from 18 (=1 MB) to 14
(=1/16th MB) fixes this. This makes it seem likely that the errors are
caused by address space fragmentation. Unfortunately, we have not yet been
able to find the root cause for this. So we use this workaround.