Disallow deletion of ethereum properties
This is for webcompat with MetaMask.

The https://wallet.polygon.technology/ site does this when web3 is
defined:
```
void 0 !== n.ethereumProvider ?
  i = n.ethereumProvider :
  void 0 !== n.web3 &&
  n.web3.currentProvider &&
  (
    n.web3.currentProvider.sendAsync &&
    (n.web3.currentProvider.send = n.web3.currentProvider.sendAsync, delete n.web3.currentProvider.sendAsync),
```

We used to not include web3 at all, then this code is not hit and
there's no problem.
Recently we started including a web3 shim so that's why it's hit now.

Now that it's hit, we were failing because the Polygon site deleted
`window.ethereum.sendAsync` and then tried to use it.
We allowed this property deletion but MetaMask did not.

Instead of the fix in this commit (with the proxy method), I wanted to instead do this in `components/brave_wallet/renderer/brave_wallet_js_handler.cc`
but that won't work because it makes delete return false.

```
- ->Set(context, gin::StringToSymbol(isolate, "networkVersion"),
+ ->DefineOwnProperty(context, gin::StringToSymbol(isolate, "networkVersion", DontDelete),
        gin::StringToV8(isolate, std::to_string(networkVersion)))
         .Check();
```

The Polygon site won't load at all if it thinks it can't delete the
property.