[network-audit] Add test coverage for the New Tab Page
As reported in https://github.com/brave/brave-browser/issues/20501,
the NTP was making outbound requests to ftx.com at startup without
user opt-in, and the network-audit didn't catch that case because
it was NOT loading a NTP as part of the different checks it does
at BraveNetworkAuditTest.BasicTests.

This patch adds explicit test coverage for that case by loading
brave://newtab and waiting 5 minutes, like with the other checks.

Resolves https://github.com/brave/brave-browser/issues/20504