Ignore chrono crate advisories again. (#11890)
Restore advisories against `chrono` and its use of the `time` 0.1
crate to the ignore list. RUSTSEC-2020-0071 and RUSTSEC-2020-0159
still have not been addressed by the upstream maintainer. We had
worked around this by porting previous uses to the newer `time` 0.3
crate as a replacement, but the SKU service and upcoming rss feed
support patch both re-introduce this dependency.

After auditing the new use to determine users are not affected, it's
best to silence the warnings so any other new issues stand out in an
`npm run audit_deps` report.

Reopens brave/brave-browser#18835