[MacOS] Apply upstream patch to filter dsymutil output.
This is an upstream patch
https://chromium.googlesource.com/chromium/src/+/9dc9fa84434a1d13450a8612a199fa26ce0a364a

This superceeds a prior temporary fix of setting use_ifc flag to false
for non official builds.

----------

Building Release on MacOS produces a huge number of linker warnings of the sort:

warning: (x86_64)  could not find object file symbol for symbol ...

This has been reported to Chromium in
https://bugs.chromium.org/p/chromium/issues/detail?id=1284576

where it was determined that the culprit is the use_icf flag.

The fix for this issue is tracked upstream in
https://bugs.chromium.org/p/chromium/issues/detail?id=1279639