Adjust install_static::GetChannelName override.
chrome/install_static/install_util.cc now uses the function
GetChannelName internally, so we need to add the _ChromiumImpl name to
the header as well. It's ok that the _ChromiumImpl version is called
internally because it only wants to know if this is the stable channel
and we only care about nightly/canary in our override.

Chromium change:

https://source.chromium.org/chromium/chromium/src/+/03039e6fc5287c0b72cb17f31f941052c2c6e0ee

commit 03039e6fc5287c0b72cb17f31f941052c2c6e0ee
Author: hmare <hmare@google.com>
Date:   Tue Jan 4 15:04:23 2022 +0000

    Add a check to validate the DmServerUrl for stable channel.

    This is validate that stable channels call the rotate command
    with a prod dmserverurl only. All other channels (dev,canary,etc)
    don't have to match this requirement.

    Bug: b/203542249