Updated TileFavicon sizes in "New Tab"
I'm proposing this change after see some favicons as "pixelated" in "New Tab" from Brave. 
After inspect it, I've realized that the size of the rendered favicon is 64px, but the style is forcing to scale up to 70px, giving the "pixelated" aspect.
I know that Brave use the chrome directory to get the visited favicons (as mentioned in brave-browser/#14485), and some favicons will be small no matter what, but I think this will reduce the impact.