Also replace the scheme with Chrome's one for brave://sync URLs
The CL below breaks BraveContentBrowserClientTest.RewriteChromeSync
since the "expected URL" we get after the rewriting process, which
still keeps the brave:// scheme, won't match the actual URL returned
by NavigationRequest::GetOriginalRequestURL(), which uses the chrome://
scheme instead. This causes the test to time out because the call to
ui_test_utils::NavigateToURL() won't ever finish as expected because of
TestNavigationObserver::OnDidFinishNavigation() never setting the value
of |last_navigation_matches_filter| to True, which happens because of
TestNavigationObserver::DoesNavigationMatchExpectedInitialUrl() will
never return True now.

A simple fix for this, which does not change production behaviour at
all, is to make sure we also rewrite the brave:// scheme with Chrome's
in BraveContentBrowserClient::HandleURLOverrideRewrite().

Chromium change:

https://chromium.googlesource.com/chromium/src/+/8a8d24ab9c59a0e12088183ade3a40c94d753d45

commit 8a8d24ab9c59a0e12088183ade3a40c94d753d45
Author: Alex Kalugin <akalugin@yandex-team.ru>
Date:   Thu Jan 20 09:21:11 2022 +0000

    Reland: Make DoesNavigationMatchExpectedInitialUrl actually check expected initial url.

    This is a reland of Ifc57f7694fe61030505e3258175546ad64144f1b

    Fix InjectSubframeDuringPendingCrossBrowsingInstanceNavigation test flakiness.

    Original change's description:

    > In https://crrev.com/c/3169962 was introduced new method
    > DoesNavigationMatchExpectedInitialUrl to TestNavigationObserver which
    > is checked in OnDidStartNavigation and OnFinishStartNavigation
    > methods. Actually currently it does not check expected_initial_url_.
    > This can cause tests flakiness (problem found researching
    > IsolatedAppTest.CrossProcessClientRedirect test flakiness).

    Bug: 1287983