[WIP] Revert back to using Clang 14 (vs Clang 15) for 256-bit integers
Clang complains with an "unsigned _BitInt of bit sizes greater than"
"128 not supported" error since the latest LLVM rollout, caused by
the change from https://reviews.llvm.org/D117238 (which got reverted
but then relanded with fixes in [1]).

As mentioned in [2], this seems to be a temporary change why the
backend-related bug is not fixed, but we need to have 256-bit integers
in the meantime for Brave Wallet, so we're temporarily reverting to
the latest version of clang 14 used in Chromium to keep building.

TODO: Find a better solution that does not force us to use Clang 14

[1] https://reviews.llvm.org/D117238
[2] https://reviews.llvm.org/D117238#3270125

Chromium changes:

https://source.chromium.org/chromium/chromium/src/+/65cf5c874dc26a5a93f5ed901c012921dca45fb8
https://source.chromium.org/chromium/chromium/src/+/cb6f83b421c6ed3c5883e7afcdeac0488b79f71f

commit 65cf5c874dc26a5a93f5ed901c012921dca45fb8
Author: Arthur Eubanks <aeubanks@google.com>
Date:   Mon Feb 14 14:45:36 2022 +0000

    Roll clang llvmorg-14-init-18258-g9477a308-1 : llvmorg-15-init-1464-gbf7f8d6f-1

    Note that this bumps the clang version from 14.0 to 15.0.

    Ran `./tools/clang/scripts/upload_revision.py bf7f8d6fa6f460bf0a16ffec319cd71592216bf4`.

    Bug: 1293292, 1296986

commit cb6f83b421c6ed3c5883e7afcdeac0488b79f71f
Author: Zequan Wu <zequanwu@google.com>
Date:   Thu Feb 10 14:00:17 2022 +0000

    Roll clang llvmorg-14-init-17086-g38e16e1c-4 : llvmorg-14-init-18258-g9477a308-1

    Ran `tools/clang/scripts/upload_revision.py 9477a308ca6b8536044c9d31a8c89062b514431e`.

    Bug: 1291928