P3A: Don't report Brave.Rewards.EnabledDuration
This metric is buggy and is being removed in favor of a better-defined
metric focussed on how long ads were enabled. As such collecting the
data is no longer neccessary.

This disabled sending the report while leaving the measurement code
in-tree so it's easier to repurpose for the ads-focussed question.

Resolves brave/brave-browser#21232