Don't show brave search for tor entries from settings
When brave search for tor is set for default provider in tor profile,
it's also added in the default provider list and it's visible from settings.

Showing brave search for tor in default search provider list makes
confusion. User would not use it for normal profile.

and onion url is set as a keyword to prevent conflict with existing
brave search entry.