Don't report refcode with P3A messages.
Always rewrite the `refcode` metadata field attached to P3A reports
to "none". We no longer need this analysis column for P3A questions
so we shouldn't collect it.

We continue to report a value as part of the metadata string to
avoid imposing a format change on the analysis back-end code.