Report US country code for all P3A messages.
When the refcode field was removed, special-case logic which allowed
the reporting the "US" country code as part of P3A message metadata
was unintentionally removed.

Since the US has many users, it's useful to segment usage data by
that region, and there is no privacy advantage to stripping that
country code.

Follow up to #12478