Replace l10n_util::GetStringUTF16 with LoadLocalizedResourceString
The strings for resources IDs requested for Brave are part of a
compressed bundle, so we need to use ui::ResourceBundle APIs or
we'll crash on startup on dcheck-enabled builds now.

Add new brave_l10n::GetLocalizedResourceUTF16String() util method as
a wrapper for ui::ResourceBundle::LoadLocalizedResourceString() that
returns a std::u16string instead of a std::string, and use it to
migrate all those calls to l10n_util::GetStringUTF16().

Chromium changes:

https://source.chromium.org/chromium/chromium/src/+/bd84936ab42d784c970b124738b97e6f5c12653a

commit bd84936ab42d784c970b124738b97e6f5c12653a
Author: S. Ganesh <ganesh@chromium.org>
Date:   Tue Mar 1 15:56:15 2022 +0000

    Add DCHECK for compressed strings to ResourceBundle::GetLocalizedString

    This is to avoid situations where `ResourceBundle::GetLocalizedString`
    is being used incorrectly instead of an API that understands and
    correctly decompresses the string, such as
    `ui::ResourceBundle::LoadDataResourceString`

    Bug: 1296797