Add support for same-directory mojo imports
This adds initial support for `import` within mojom files. Due to limitations with where the generated files are written too and that we cannot use full paths within the the headers because generated headers are copied into a flat Headers directory within the compiled framework. With those things in mind, there are two restrictions in place:

1. Only imports of Brave mojom files are considered, base mojom ones are handled manually via custom typemaps within the generator code
2. All `sources` in a `mojom` target must exist in the same directory.