[MacOS] Copy provisionprofile with identifier name.
Chromium change:

https://chromium.googlesource.com/chromium/src/+/1119a66c69e4c5f8671192a56f59af71b65d9ab1

commit 1119a66c69e4c5f8671192a56f59af71b65d9ab1
Author: Avi Drissman avi@chromium.org
Date: Fri Mar 4 19:12:47 2022 +0000

Attach the SHA-1 hash of the signing identity to the provisioning profile filename

Provisioning profiles embed a list of certificates that can sign the
app in which they are embedded. We're rolling the certificate that
is being used to sign Chrome, and we have new provisioning profiles
for use with that new certificate, so teach the signing scripts to use
the provisioning profile that matches the certificate being used.

(cherry picked from commit b2ffdf63dd539b7ba00202b435153f6bd63901aa)

Bug: 1299959