Disables upstream unit tests due to Greaselion donwload service.
* InstalledExtensionMenuItemViewTest.*.
* SidePanelCoordinatorTest.*
* SitePermissionsHelperUnitTest.*

These tests crash due to
TestingBraveBrowserProcess::greaselion_download_service() having
NOTREACHED which causes immediate crash.

TODO: The code in GreaselionServiceFactory::BuildServiceInstanceFor:

  // Brave browser process may be null if we are being created within a unit
  // test.
  if (g_brave_browser_process)
    download_service = g_brave_browser_process->greaselion_download_service();

is not sufficient because g_brave_browser_process is not null in
these unit tests and download_service is then passed to c'tor and is
expected not to be null.