P3A: Move metric list to a separate file.
Adding new P3A probe requires updating the list of allowed metrics
so the new measurement can be reported. This is a lower review bar
than changing the reporting code itself, so we'd like to have it
in a different place so we can have a separate list of code owners
for it.

We use a base::flat_set for quick membership testing in
BraveP3AService::IsActualMetric(). Instead of creating
a local static for this, replace the underlying array
with a `constexpr fixed_flat_set` so the metric list can
be sorted at compile time and we don't need the messy
`base::NoDestructor` decoration.

Thanks to @iefremov for the suggestion.

kCollectedHistograms is marked with the C++17 `inline` attribute,
which should ask the linker to remove any duplicate copies of the
symbol, minimizing the code size contribution of having the allow
list in a separate file.

Resolves https://github.com/brave/brave-browser/issues/22263