[Android] New c-tor signatures for top toolbar classes
Chromium change:
https://github.com/chromium/chromium/commit/c93a4543bd360aa2ca31f2bccee6bd1381427eb8

AboutThisSite: Launch ephemeral tabs from PageInfo
Pass EphemeralTabCoordinator to PageInfo in order to open source info
and the "more about this page" link on top of the current tab instead
of in a new tab.
Add EphemeralTabCoordinatorSupplier to RootUiCoordinator to make it
available to all the places that create PageInfo.

Screen recording: https://crbug.com/1318000#c5

Bug: 1318000