Disable onion redirect pref along with Tor.
Update the "Automatically redirect .onion sites" preference to
track the new form of the "Private window with Tor" pref, which
was changed to address a console warning.

Leaving this pref iteractive when Tor is disabled is confusing
because it implies onion-alternate sites will continue to redirect
when this is not the case.

Resolves https://github.com/brave/brave-browser/issues/23560