Move Solana Ledger integration to untrusted iframe
The Solana Ledger JS libraries are now accessed by the chrome://wallet
and chrome://wallet-panel contexts via postMessage calls to an iframe
whose origin is chrome-untrusted://ledger-bridge.

If we detect that user needs to grant the Wallet permssion to access a
Ledger device, an 'Authorize' button is displayed *within* the untrusted
frame so that the permission dialog (which requires a user gesture to be
displayed) screen is shown.

When permission is granted, the chrome-untrusted://ledger-bridge sends a
postMessage to the parent (trusted) context instructing it to continue the flow.