update CSP to minimize available scripts that can be loaded
Since we were only ever using load_time_data.js the CSP here
can be updated so that only that file can be loaded. Additionally,
since CR/1173575 deprecates the *.js files in favor of *.m.js
files, I've updated this to the newer file so that it won't throw
the warning anymore.