[CodeHealth] Broken code flagged by clang-tidy
This change is the result of running the following clang-tidy
modernisers:

  bugprone-argument-comment
  bugprone-assert-side-effect
  bugprone-dangling-handle
  bugprone-inaccurate-erase
  bugprone-string-constructor
  bugprone-string-integer-assignment
  bugprone-unused-raii
  bugprone-use-after-move

There are actual UB, and crashes that have been caught specially in the
user-after-move category. One bug was also found on the use of
innaccurate erasure. A similar bug to this one had caused a crash to the
nightly release this week.