Re-enabling -Wunreachable-code-aggressive
This warning flag was being overriden due to build errors on our
overrides with chromium. However, this build warning is very useful, and
we can suppress it with `if ((true))` statements that will elide the
code.

This warning can be very useful in pointing out small misconceptions
in code behaviour.