Moved SidebarContainerView to contents_container
So far we used another parents view for holding upstream's
contents_container and sidebar container view. And this new parent
view is set as new contents_container in BrowserView.
However, this breaks BrowserView's some assumptions that
devtools_webview or contents_webview were direct children of
contents_container. To fix this, SidebarContainerView is moved
to contents_container and it's layout is done via subclass of
ContentsContainer.