[CodeHealth] Use base::ranges::find* algorithms
Using ranged algorithms leads to slightly simpler code. It is also
better communicates the semantics involved. For cases where `std::find`
was merely being used to confirm that an element was present in a
container, `base::Contains` has been used, as it describes better what
is taking place.

Additionally, this change removes a lot of cases where a projection
suffices in place of a lambda, simplifying the code altogether.