Implement eth_coinbase handling for MM compat
This returns the same information as eth_accounts. No extra information
is leaked (e.g. with regards to potential fingerprinting data).

The main difference is that it returns only 1 string account or null, instead of
an array with 0 or 1 item.  This is needed for webcompat reasons for
sites like bsc.cryptoz.cards which use eth_coinbase instead of
eth_accounts.