[CodeHealth] Removing uses of PrefService::Get
Due to the ongoing base::Value modernisations going on upstream,
PrefServices::Get has been deleted already on the main branch, in favour
of type specific versions, like PrefService::GetValueDict.

This change removes all uses of PrefService::Get that are possible to be
removed for now, to reduce the disruptions in upcoming version upgrades.