Make Dapp addresses lowercase to be compatible with MM
Fix https://github.com/brave/brave-browser/issues/25039

Adobe connect MetaMask feature doesn't work for this URL: https://connected-accounts.adobe.com/

This turns out to be because the accounts we return EIP-55 checksumed addresses to Dapps instead of lowercase like MetaMask.

On this particular Adobe site, the connected account is sent up in a JSON structure to Adobe with a POST HTTPS request to
https://connected-accounts.adobe.com/api/connection/metamask.
If that address is not lowercase, the server responds with an HTTP status code of 400.
The Adobe site uses the address that is given from the `eth_requestAccounts` call they make.

MM has lowercase addresses for return info via these functions: `eth_requestAccounts`, `eth_accounts`, `eth_coinbase`, `wallet_requestPermissions`, and `wallet_getPermissions`
For the best possible webcompat we need to unfortunately follow their standard in the wild.
This pull requests makes the return values to Dapps lowercase.