[Rewards] Do not start utility process until user opted in.
This change splits onboarding UI out of the settings so that we can show
onboarding without starting the utility process needed for the settings.

It also splits out the tour UI code because it's used in both onboarding
and settings.

Additionally, in the rewards service StartLedgerIfNecessary only starts
the utility process when rewards are "enabled". Creating a wallet starts
the utility process unconditionally as this is now the equivaling of
opting in.