Disable tab cycling when 'vertical tabs' is enabled
On Linux, there's a feature once called 'tab cycling', which switches
the active tab on mouse wheel events. But this doesn't go well with
vertical tabs as they are in scroll view. Also, upstream has disabled
this feature when scrollable tabstrip is enabled. In favor of this,
disable tab cycling when vertical tabstrip is enabled.