[Android] Java code adjustments to the upstream changes
Chromium change:
https://github.com/chromium/chromium/commit/2b70a2044f6ac7f362f2730fef52134ca5d68083

[M107 Merge][WebContentsImpl] Don't call didFinishNavigation in child…
… frames

Part of reducing navigation cost to reduce scroll jank. Java observers
don't do work when didFinishNavigation is called in child frames,
so we don't have to notify java in this case.

The changes to restrict didFinishNavigation to main frames is not behind a flag, so to measure the effect on jank, we added a didFinishNavigationNoop function (behind a flag) to mirror what didFinishNavigation did before this change. We think that the overhead from overrides add up to something measurable, so I added a redundant boolean check to make sure this override doesn't get optimized (might get optimized if it's empty but not completely sure). I also wanted it to be as close as possible to what it was doing before, so that we can capture more accurately the effect on jank metrics.

(cherry picked from commit df8db0d)

Bug: 1359110, 1369683