Support overlay mode for vertical tab strip
Now we have 3 states for vertical tab strip.
* Collapsed: tab strip will be the minimum size
* Expanded: tab strip will be the maximum size and contents area will be
             moved next to it.
* Floating: tab strip will be the maximum size but it'll cover some of
            contents area.

Floating state will be triggered by mouse hover state and Expanded state
will be triggered by a button on header.

Note that we need to embed VerticalTabStripRegion on another child widget.
This is because when we have a view atop web contents area, web contents
will precede the tab strip in terms of event handling. In order to handle
mouse events ahead of web contents, we need to embed the tab strip on
a widget.