Don't take vertical tab strip into account when calculating the min window size
This is a trial to fix a bug, where window grows too tall and becomes
unresizable.

The exact steps to reproduce was not found, but it seems somehow the
min height is set to window. As min size could be cached internally,
this could have happen when the min size isn't flushed.