Remove top shadow for omnibox popup when using vertical tabs and hiding window title (#15819)
* Remove top shadow for omnibox popup when using vertical tabs and hiding window title

In this case, popup's bounds shouldn't be out of browser. Especially on
Mac, Widgets can't be out of workspace bounds. If we set a coordinate
that out of the workspace, it'll be adjusted by the system. This leads
to UI glitch. So we should remove top shadow of the popup so that it
can stay inside browser.