Fixed sidebar's horizontal alignment is not properly set at startup
fix https://github.com/brave/brave-browser/issues/26605

Set horizontal alignment when sidebar initialization is completed.

When SidebarContainerView::SetSidebarOnLeft() is called before sidebar
is initialized, it just update |sidebar_on_left_| and it's not propagated
to children because they're not yet initialized.
After initialized, it was called but it's early returned because
|sidebar_on_left_| was set already. This bug came frome here.