Reset window controls position when NC inset changes on Mac
Previously on Mac, window control's y coordinate wasn't adjusted when
options changes, even though the options affect the non-client area's
height. In order to solve this, make a call to undocumented method of
NSThemeFrame: _resetButtons when it's needed.

Also, adjust window controls' y position when title is hidden with
vertical tab strip so that it could be aligned with toolbar's height