Fix brave/brave-ios#6436: Update RPC Service extensions for use with any network (brave/brave-ios#6437)
Update RPC Service extensions to add function to fetch all networks for our supported coin types, and update the balance functions to add network as a parameter. Portfolio will be updated to display assets from all networks, so we need to be able to fetch the balance for specific networks other than the currently selected network.