Fix brave/brave-ios#6623, fix brave/brave-ios#6633: Script injection overrides and double injected scripts (brave/brave-ios#6631)
Scripts injected into ALL frames must only be included once in that frame. Otherwise the script can throw an exception if the page modifies its source code at runtime and WebKit attempts to inject it a second time even though it is already injected.

Do NOT `freeze` `toString` on custom objects we don't own.
Scripts and objects that provide a custom `toString` should freeze it themselves.