[MacOS] Fix signing of privileged helper.
Privileged helper name is hardcoded in the branding file right now to
start with org.chromium.Chromium. On the other hand the signing script
uses the proper branding name (com.brave.Brave<.CHANNEL>), which causes
a mismatch and a signing error.

If we eventually start using the updater we should change the branding
appropriately (including the channel), but for now we can update the
signing script with the hardcoded name.

ALSO, make sure the the feature controlling the use of the updater is
disabled.

Chromium change:

https://chromium.googlesource.com/chromium/src/+/3b744610ade7a3cf766f215d78f773e93e76341c

commit 3b744610ade7a3cf766f215d78f773e93e76341c
Author: Joshua Pawlicki <waffles@chromium.org>
Date:   Fri Oct 21 17:01:44 2022 +0000

    Updater: Mac: Embed Chromium Updater in Chrome, behind Finch.

    This effectively replaces compile-time switches with run-time switches.

    It also renames the privileged helper binary and sets up signing for it.

    This also fixes a compilation failure in a unit test that was broken a
    long time ago and not run. Unfortunately, the test does not pass at ToT
    even with the compile fix and so it is disabled to be handled in a
    later CL.

    Although it now has no effect, I didn't actually remove the
    enable_chromium_updater buildflag, since it is still set in some builder
    configs, and I'd like to transition it into enable_updater.

    This causes a 6-12% increase in Chrome's DMG size.

    Fixed: 1370532