Split pinned tabs and non-pinned tabs area for vertical tab strip (#16165)
* Split pinned tabs and non-pinned tabs area for vertical tab strip

Previously, we can't tell unpinned tabs from pinned tabs as they
look alike. In order to solve this, we need to change layout constraints
of pinned tabs - Use grid layout for pinned tabs.

In order to do this, make use of CompoundTabContainer from upstream.
CompoundTabContainer consists of two TabContainerImpl. One is for pinned
tabs and the other is for unpinned tabs. With this we can easily
distinguish pinned tabs and unpinned tabs. In addition, we have another
feature request to make unpinned tabs sticky, which means pinned tabs
should be out of scroll area. Thus, using CompoundTabContainer would be
a good move for that as well.