Fix crash on drag-and-drop (#16502)
When dropping tabs to create a new window, the browser crashes.
This happens because we're not updating TabContainer properly.

During the drag-and-drop session, TabSlotViews are heavily bound to
TabDragController and TabDragContext. So we should reuse the existing
views as much as possible and synchronize with the previous state, when
we update tab containers.