Let CI distinguish test compilation and test run failures (#15600)
Previously, `npm run test --output` exited with a non-zero code on both
compilation and test errors. This made it impossible to determine which
of the two happened - and thus hard to decide whether to generate test
reports. Now, the command exits with 0 on test failures. (Note that the
behavior is unchanged when --output is not given. That is, in this case
you still get non-zero exit codes for test failures.)