Farble window.{resizeTo,moveTo}; enforce minimum farbled screen size
When farbling screen size, make sure window.resizeTo and window.moveTo
use the farbled coordinate system.

When a window.open(...) popup is opened with very small dimensions,
scripts should not be told that the screen is extremely small. So
we put a minimum screen width and height, with a random number
between 0 and 8 added to each dimension.

Fixes #27492