Brave News Location Bar Button gets updated more accurately
The button is now a PageActionIconView so we can call Update() more frequently and use a delegate to check if it should be shown, so that the button cannot lose the state of whether page actions should be shown.
Also fixes InkDrop to stay highlighted when bubble is showing.
And height is more accurately controlled via the LocationBar.