[iOS] Load cpp type mappings info for correct move only/copyable codegen
This loads base mojom type mappings (url/origin/time/etc.) and updates the codegen for Obj-C wrappers to use the correct definition of said mappings in C++ function definitions.

This fixes an issue where if a GURL was referenced in a interface functions return type it would fail to generate a buildable Obj-C wrapper because GURL would be used instead of const GURL& in a param context.