Use a generic cxx version reference.
We want to control the version of cxx we use from the top-level
build/rust/Cargo.{toml|lock} so reference a generic "1.0" version
in the individual crates instead of trying to reference the correct
vendored copy directly with a relative path, which is error-prone
and harder to review.

This reverts us from the the chromium-vendored 1.0.81 to our own
vendored 1.0.68.

Building with 1.0.81 has been working fine. However, we use the
`cxxbridge` command-line tool to generate related binding files
are part of the build, preferring this to the less-controlled
`build.rs` pattern also supported by the cxx crate. Our setup
scripts try to install a matching version of the tool to avoid
any compatibility issues. Unfortunately, while the current releases
of the crate itself build fine, recent versions of the cxxbridge
tool require clap 4.0, which has a minimum-supported rust toolchain
version of 1.60, which is newer than we currently support.

Between the two choices of splitting the crate and tool versions
or using our older vendored copy, this commit chooses the later.