Get metadata from m.youtube.com pages (#16419)
* Get title and thumbnail from m.youtube.com pages

Youtube dynamically updates its contents and title. So The title that we
get is
fixed as "Youtube" which is set before the actual video's title is set.

Also, they don't have og:image tag, which we were looking up as
thumbnail.

So in thi patch, get these data from youtube's internal data. Could be
quite fragile.