[Android] Consolidate v8 asset targets in //gin
Chromium change:
https://github.com/chromium/chromium/commit/22ff125de8f82dd3cfaa6737b7726be2c82180c3

Before:
* Primary ABI assets were taken from //v8
  * Except when use_v8_context_snapshot=true they were taken
    from //tools/v8_context_snapshot.
* Secondary ABI assets were taken from //android_webview

After:
* Both primary and secondary abi asset targets live in //gin.
  * //gin/v8_initializer.cc is what cares about the snapshot
    filenames, so this package makes the most sense to hold the
    android_asset targets.

This removes duplication and makes it easier to refactor in future
(e.g. this is a step towards using narrower deps for the snapshots,
which I will do in a follow-up).

This also removes assets from a couple of source_set targets, as
they had no effect there.

Bug: 1402705