[Android] Refactor ExternalNaviagation Tab Clobbering
Chromium change:
https://github.com/chromium/chromium/commit/5174b2669233382244250f7f1e1352303dbcb2c0

Refactor ExternalNaviagation Tab Clobbering into InterceptNavigationImpl
No functional changes (outside of an edge cases where we forgot to call
the async action callback and missed clearing history for a clobbered
tab in the case of a link to another browser).

This change moves the tab clobbering logic into
InterceptNavigationDelegateImpl, and also makes the async action
callback mandatory, so it can't be mistakenly missed in the future.
This will allow us to implement fallback URL support for subframes.

Bug: 1365100