Revert to cargo-audit 0.16.0
Pin the version of `cargo-audit` we install and specify the
`--locked` option to use the originally published versions
of all dependencies instead of upgrading to the latest semver-
compatible releases.

Recent versions of `miniz_oxide` and `toml` raise the
minimum-supported rust version past the 1.59.0 we're
currently supporting. We can't bump that because of
macos linker issues with eh_personality routines, so
stay on an old version for now. None of the 0.17.x releases
currently build for us.